backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.5.3
numpy==2.4.3
openai==2.28.0
packaging==26.0
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
python-dotenv==1.0.1
